backend/requirements.txt
fastapi
uvicorn[standard]
pydub
pyloudnorm
librosa
numpy
soundfile
celery==5.3.*
redis>=4.5
python-multipart
pydantic-settings
piper
pytest
httpx
